+++ ABCRetailers.Functions/Functions/OrdersFunctions.cs
@@ -106,8 +106,8 @@ public class OrderMessage
     public string ProductId { get; set; }
     public string ProductName { get; set; }
     public int Quantity { get; set; }
-    public decimal UnitPrice { get; set; }
-    public decimal TotalPrice { get; set; }
+    public double UnitPrice { get; set; }
+    public double TotalPrice { get; set; }
 }
 
 public class UpdateOrderStatusRequest
